Lock Free Stack #4
Using Counters for Implementing Lock Free Stack.
Similar to the shared_ptr mechanism but this approach has 2 unique access tracking values.

externalCount starts with 1 and keeps increasing when it requests TryPop().
It controls the Referencing rights of the Node.

the internalCount is the atomic type and starts with 0.
Add value when one thread acquires Ownership of the Node in TryPop().
Sub value when other threads that didn't make it out to acquire Ownership of the Node in TryPop().
It works like a reference counter of the Node.