Lock Free Queue #2
Using Counters for Implementing Lock Free Queue.

It needs to have secure the head and tail for the multi-threaded push & pop scenario.

NodeCounter has an externalCountRemaining variable that checks the node is being used in both push & pop.

If the externalCountRemaining is 0. there's no more race.
It is controlled by FreeExternalCount() to decrease it.